radicle-term: Remove custom pager
I learned that there is an implementation of a pager in
`crates/radicle-term/src/pager.rs`. It does not make sense to me to
maintain a pager in `heartwood`, this is too far from the core of what
we are trying to achieve and adds unnecessary maintenance burden. We can
allow ourselves to defer to the paging done by `git` or just call the
user's `$PAGER`. Even though falling back to unpaged output might not be
the nicest thing, I still prefer that over maintaining the pager.

It also was frustrating to me that the pager does not support things
like search (like `less` does).

This also significantly decreases our dependence on `termion`.

The pager was originally implemented by cloudhead in

  40f4383bbc94acbcadd06fce41e2960a6ffb9c33

and integrated into `rad diff` in

  6dd52c94fb1472e251110e439fe656aaed74019f

This commit restores the behaviour introduced in

  84a95848b02de5599f5a2dc905fa2ba4c185b7ab

(before there was a custom pager implementation).